requirements.txt
annotated-types==0.6.0
anyio==4.2.0
attrs==23.1.0
Automat==22.10.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
constantly==23.10.4
contourpy==1.2.0
cryptography==41.0.7
cssselect==1.2.0
cssselect2==0.7.0
cycler==0.12.1
defusedxml==0.7.1
distro==1.8.0
easyocr==1.7.1
filelock==3.13.1
fire==0.5.0
fonttools==4.47.0
fsspec==2023.12.2
fugashi==1.3.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
h5py==3.10.0
hpack==3.0.0
hstspreload==2023.1.1
html5lib==1.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.20.1
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
imageio==2.33.1
importlib-metadata==7.0.1
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
jaconv==0.3.4
Jinja2==3.1.2
jmespath==1.0.1
kiwisolver==1.4.5
langdetect==1.0.9
lazy_loader==0.3
loguru==0.7.2
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib==3.8.2
mplcursors==0.5.2
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.2
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
packaging==23.2
parsel==1.8.1
pathspec==0.12.1
Pillow==10.1.0
Protego==0.3.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
pydantic==2.5.2
pydantic_core==2.14.5
PyDispatcher==2.0.7
pydyf==0.8.0
PyMuPDF==1.23.7
PyMuPDFb==1.23.7
pyOpenSSL==23.3.0
pyparsing==3.1.1
PyPDF2==1.26.0
pypdfium2==4.25.0
pyperclip==1.8.2
pyphen==0.14.0
pytesseract==0.3.10
python-bidi==0.4.2
python-dateutil==2.8.2
PyYAML==6.0.1
queuelib==1.6.2
rapidfuzz==3.6.0
regex==2023.10.3
reportlab==4.0.8
requests==2.31.0
requests-file==1.5.1
rfc3986==1.5.0
safetensors==0.4.1
scikit-image==0.22.0
scipy==1.11.4
service-identity==23.1.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
sympy==1.12
termcolor==2.4.0
tesseract==0.1.3
tifffile==2023.12.9
tinycss2==1.2.1
tldextract==5.1.1
tokenizers==0.15.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.66.1
Twisted==22.10.0
typing_extensions==4.9.0
Unidecode==1.3.7
unidic-lite==1.0.8
urllib3==2.1.0
w3lib==2.1.2
weasyprint==60.2
webencodings==0.5.1
zipp==3.17.0
zope.interface==6.1
zopfli==0.2.3
